added more WIP commands
